+++ src/application/context/components/remove/ComponentRemovedEventHandler.ts
@@ -2,29 +2,22 @@ import { IEventHandler } from "../../../messaging/IEventHandler.js";
 import { BaseEvent } from "../../../../domain/BaseEvent.js";
 import { ComponentRemovedEvent } from "../../../../domain/components/remove/ComponentRemovedEvent.js";
 import { IComponentRemovedProjector } from "./IComponentRemovedProjector.js";
-import { RelationDeactivationCascade } from "../../relations/deactivate/RelationDeactivationCascade.js";
-import { IRelationMaintenanceGoalRegistrar } from "../../relations/maintain/IRelationMaintenanceGoalRegistrar.js";
+import { RelationPruningCascade } from "../../relations/prune/RelationPruningCascade.js";
 import { EntityType } from "../../../../domain/relations/Constants.js";
 
 export class ComponentRemovedEventHandler implements IEventHandler {
   constructor(
     private readonly projector: IComponentRemovedProjector,
-    private readonly relationDeactivationCascade: RelationDeactivationCascade,
-    private readonly relationMaintenanceGoalRegistrar: IRelationMaintenanceGoalRegistrar
+    private readonly relationPruningCascade: RelationPruningCascade
   ) {}
 
   async handle(event: BaseEvent): Promise<void> {
     const componentRemovedEvent = event as ComponentRemovedEvent;
     await this.projector.applyComponentRemoved(componentRemovedEvent);
-    await this.relationMaintenanceGoalRegistrar.execute(
+    await this.relationPruningCascade.execute(
       EntityType.COMPONENT,
       componentRemovedEvent.aggregateId,
-      "component was removed"
-    );
-    await this.relationDeactivationCascade.execute(
-      EntityType.COMPONENT,
-      componentRemovedEvent.aggregateId,
-      `Automatically deactivated because component ${componentRemovedEvent.aggregateId} was removed`
+      `Automatically pruned because component ${componentRemovedEvent.aggregateId} was removed`
     );
   }
 }